[compiler][wip] Remove old mutation/aliasing implementation
